requirements.txt
aiofiles==24.1.0
aiogram==3.20.0.post0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
async-timeout==5.0.1
asyncio==3.4.3
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
chart-studio==1.1.0
click==8.1.8
distro==1.9.0
dnspython==2.7.0
exceptiongroup==1.2.2
Flask==3.1.0
frozenlist==1.6.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
loguru==0.7.3
magic-filter==1.0.12
MarkupSafe==3.0.2
motor==3.7.0
multidict==6.4.3
narwhals==1.35.0
nowpayments==1.3.3
openai==1.75.0
packaging==25.0
pillow==11.2.1
plotly==6.0.1
propcache==0.3.1
psycopg2-binary==2.9.10
pyaes==1.6.1
pyasn1==0.6.1
pycoingecko==3.2.0
pycountry==24.6.1
pycryptodome==3.22.0
pydantic==2.11.3
pydantic_core==2.33.1
PyJWT==2.10.1
pymongo==4.12.0
Pyrogram==2.0.106
pyshorteners==1.0.1
PySocks==1.7.1
python-dotenv==1.1.0
python-telegram-bot==22.0
pytz==2025.2
quickchart-io==2.0.0
requests==2.32.3
retrying==1.3.4
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
stripe==12.0.1
Telethon==1.40.0
TgCrypto==1.2.5
tornado==6.4.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzlocal==5.3.1
urllib3==2.4.0
Werkzeug==3.1.3
yarl==1.20.0
